requirements.txt
# Core Dependencies
requests>=2.31.0

# Excel and Office Document Generation
openpyxl>=3.1.2

# PDF Generation
reportlab>=4.0.7

# PowerPoint Generation
python-pptx>=0.6.23

# Word Document Generation
python-docx>=1.1.0

# Environment Variables
python-dotenv>=1.0.0
